+++ HISTORY.md
@@ -49,6 +49,9 @@
     * German
     * Chinese, thanks, Simon!
 
+### 2021-11-08
+* objects: renamed some internal color methods
+
 ### 2021-11-07
 * widgets, blocks, byob: allow block-instances to be dragged off from templates in the "hide blocks" dialog
 * gui: prevent switching to another sprite if a "hide blocks" dialog is open